Move functions added - error in valgrind
realloc error in valgrind

+++ include/structs.h
@@ -23,8 +23,8 @@ typedef struct Move
     int origin;
     int destiny;
     int ton;
-    int cost;
-    
+    float cost;
+
 } Move;
